requirements.txt
Pillow==11.3.0
numpy==2.3.2
# Add any other dependencies your project needs here
